Removed runjs in favor of vanillajs scripts
